requirements.txt
O365==2.0.27
python-dotenv==1.0.0
requests==2.28.2
